# M6/ft_transcendence/backend/src/docs/swagger-output-public-api.json
{
  "swagger": "2.0",
  "info": {
    "title": "AlpacaParty Public Api",
    "description": "Auto-generated Swagger documentation.\n## Getting Started\n\n1. **Log in** to your AlpacaParty account.\n2. Go to **Profile &rarr; Settings &rarr; Public API Key** and click **Generate Key**.\n3. Copy the key. \n4. Press the **Authorize** Button, paste the key and press authorize.\n\n_You can reveal, regenerate, or revoke the key at any time from the same page._\n",
    "version": "1.0.0"
  },
  "host": "10.13.11.3:8443",
  "basePath": "/api/public",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-api-key",
      "description": "api-key for public api"
    }
  },
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "PublicApi"
        ],
        "description": "Get users",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Failed."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "PublicApi"
        ],
        "description": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the user to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Failed."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts": {
      "get": {
        "tags": [
          "PublicApi"
        ],
        "description": "Get posts",
        "responses": {
          "200": {
            "description": "Success."
          },
          "400": {
            "description": "Failed."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "PublicApi"
        ],
        "description": "Create a new post",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                },
                "imageUrl": {
                  "example": "any"
                },
                "isPublic": {
                  "example": "any"
                }
              }
            },
            "description": "Post creation data",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Post created successfully."
          },
          "400": {
            "description": "Failed validation."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/posts/{id}": {
      "put": {
        "tags": [
          "PublicApi"
        ],
        "description": "Update an existing post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the post to update"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "example": "any"
                },
                "imageUrl": {
                  "example": "any"
                },
                "isPublic": {
                  "example": "any"
                }
              }
            },
            "description": "Updated post data",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Post updated successfully."
          },
          "400": {
            "description": "Failed validation."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "PublicApi"
        ],
        "description": "Delete a post by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the post to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully."
          },
          "400": {
            "description": "Failed."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  }
}